Split Academy Ofsted methods into a separate file
